chore: utils 파일 분리
